requirements.txt
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioitertools==0.12.0
aiosignal==1.3.2
aiosmtplib==3.0.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bandit==1.8.5
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
boto3==1.37.3
botocore==1.37.3
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
coverage==7.9.1
cryptography==45.0.4
dill==0.4.0
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.4.0
fastapi==0.115.13
fastapi-cli==0.0.7
fastapi-mail==1.5.0
filelock==3.18.0
filetype==1.2.0
flake8==7.3.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.5.0
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
pbr==6.1.1
phonenumbers==9.0.7
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
psycopg2==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.7
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
python-barcode[images]==0.15.1
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
s3transfer==0.11.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
stevedore==5.4.1
tenacity==9.1.2
text-unidecode==1.3
tomli==2.2.1
tomlkit==0.13.3
typer==0.16.0
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
urllib3==2.5.0
user-agents==2.2.0
userpath==1.9.2
uvicorn==0.34.3
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
